requirements.txt
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
APScheduler @ file:///croot/apscheduler_1668632958443/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644553347904/work
bcrypt @ file:///tmp/abs_6fpe92qzzo/croots/recipe/bcrypt_1659554336150/work
cffi @ file:///croot/cffi_1670423208954/work
click @ file:///tmp/build/80754af9/click_1646056706450/work
cryptography @ file:///croot/cryptography_1689373676338/work
DBUtils==3.0.3
exceptiongroup==1.1.3
fastapi==0.101.1
h11 @ file:///tmp/build/80754af9/h11_1620423447028/work
idna==3.4
iso8601==1.1.0
passlib==1.7.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
PyMySQL @ file:///home/builder/ci_310/pymysql_1640812746452/work
pypika-tortoise==0.1.6
python-dotenv @ file:///croot/python-dotenv_1669132557191/work
python-multipart==0.0.6
pytz @ file:///croot/pytz_1671697431263/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzlocal @ file:///croot/tzlocal_1668627599277/work
uvicorn @ file:///croot/uvicorn-split_1678090074528/work
